fix: add retry logic for graphql fetch
